Mass Conversion
These are 4 powershell scripts that use the inbuilt 'save as pdf' function in microsoft office to batch process files.

The script saves the files within the same filinf structure as the original office file, and also maintains the original 'last modified' date. Failures are harvested in a csv and a timer for 30 seconds is initialised to address periods of inactivity / stuck script identification.